+++ Danki2/Assets/Scripts/Abilities/Channel/Whirlwind.cs
@@ -5,10 +5,10 @@ public class Whirlwind : Channel
 {
     private static readonly float _spinRange = 2;
     private static readonly float _spinDamageMultiplier = 0.3f;
+    private static readonly float _spinDamageInterval = 0.5f;
     private static readonly float _selfSlowMultiplier = 0.5f;
     private static readonly float _finishRange = 3;
     private static readonly float _finishDamageMultiplier = 1;
-    private static readonly float _damageTickInterval = 0.5f;
 
     private WhirlwindObject _whirlwindObject;
 
@@ -23,7 +23,7 @@ public class Whirlwind : Channel
     public override void Start()
     {
         slowEffectId = Context.Owner.AddPassiveEffect(new Slow(_selfSlowMultiplier));
-        repeater = new Repeater(_damageTickInterval, () => AOE(_spinRange, _spinDamageMultiplier));
+        repeater = new Repeater(_spinDamageInterval, () => AOE(_spinRange, _spinDamageMultiplier));
 
         Vector3 position = Context.Owner.transform.position;
         Vector3 target = Context.TargetPosition;
@@ -65,32 +65,5 @@ public class Whirlwind : Channel
                 actor.ModifyHealth(-damage);
             }
         });
-    }
-}
-
-public class Repeater
-{
-    private readonly float interval;
-    private readonly Action action;
-
-    private float currentTime = 0;
-
-    public Repeater(float interval, Action action, bool runOnStart = true)
-    {
-        this.interval = interval;
-        this.action = action;
-
-        if (runOnStart) action.Invoke();
-    }
-
-    public void Update()
-    {
-        while (currentTime >= interval)
-        {
-            action.Invoke();
-            currentTime -= interval;
-        }
-
-        currentTime += Time.deltaTime;
     }
 }
