+++ Danki2/Assets/Scripts/Utils/Repeater.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4d7ad90409ee1f48868b92baf954f60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
